# backend/docs/swagger.yml
openapi: 3.0.0
info:
  title: API
  version: 1.1.0
  description: Swagger API Documentation

servers: 
  - url: http://localhost:5000

paths:
  /comics:
    # Get all comics
    get:
      summary: Get all comics
      responses:
        '200':
          description: A list of all saved comics
    
    # Add a comic
    post:
      summary: Create a new comic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - author
                - issue
                - volume
              properties:
                title:
                  type: string
                author: 
                  type: string
                issue:
                  type: number
                  minimum: 1
                volume:
                  type: number
                  minimum: 1
                publisher:
                  type: string
                coverVariant:
                  type: string
                releaseDate:
                  type: string
                  format: date
                rating:
                  type: number
                  minimum: 1
                  maximum: 10
                purchasePrice:
                  type: number
                currentValue:
                  type: number
                condition:
                  type: string
                  enum: 
                    - Near Mint
                    - Very Fine
                    - Fine
                    - Very Good
                    - Good
                    - Fair
                    - Poor
                  default: Good
                isRead:
                  type: boolean
                  default: false
                genre:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                notes:
                  type: string
      responses:
        '201':
          description: Comic created
          
  # GET a comic by id
  /comics/{id}:
    get:
      summary: Retreive a Comic by ID
      description: Returns a single comic object based on the provided ID.
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: string
          description: The ID of the Comic to retreive
      responses:
        '200':
          description: Comic by ID
        '404':
          description: Comic not found